pyproject.toml
[tool.poetry]
name = "pictalk"
version = "0.1.0"
description = ""
authors = ["Anuj Jindal <31108317+jindalAnuj@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
transformers = "^4.40.1"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
tensorflow = "^2.16.1"
tf-keras = "^2.16.0"
langchain = "^0.1.16"
langchain-openai = "^0.1.4"
streamlit = "^1.33.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
